face_emotion/requirements.txt
numpy==1.15.1
pillow==5.2.0
requests==2.19.1
flask==1.0.2